new way of implementation and it works
