+++ sage/example-transition/src/types.rs
@@ -1,0 +1,87 @@
+//! Some example transitions.
+//!
+//! These should be expanded to really showcase the power of the SageApi design.
+
+use frame_support::sp_runtime::testing::H256;
+use parity_scale_codec::{Decode, Encode, MaxEncodedLen};
+use sage_api::AssetT;
+use scale_info::TypeInfo;
+
+pub type AssetId = H256;
+
+/// Placeholder type, this was just a quick brain dump to get things going.
+#[derive(Copy, Clone, Debug, PartialEq, Eq, Encode, Decode, MaxEncodedLen, TypeInfo)]
+pub struct Asset {
+	pub collection_id: u32,
+
+	pub asset_type: u32,
+
+	pub asset_sub_type: u32,
+
+	pub dna: [u8; 32],
+
+	pub minted_at: u32,
+
+	// Example of a game's custom field.
+	pub level: Level,
+
+	// Example of a game's custom field.
+	pub consumed: bool,
+}
+
+#[derive(Copy, Clone, Debug, PartialEq, Eq, Encode, Decode, MaxEncodedLen, TypeInfo)]
+pub enum Level {
+	One,
+	Two,
+	Three,
+	Max,
+}
+
+pub const MAX_LEVEL_REACHED_ERROR: u8 = 100;
+pub const ASSET_ALREADY_CONSUMED: u8 = 101;
+
+impl Level {
+	pub fn upgrade(self) -> Result<Self, sage_api::Error> {
+		use Level::*;
+		match self {
+			One => Ok(Two),
+			Two => Ok(Three),
+			Three => Ok(Three),
+			Max => Err(sage_api::Error::Transition { error: MAX_LEVEL_REACHED_ERROR }),
+		}
+	}
+}
+
+impl AssetT for Asset {
+	fn collection_id(&self) -> u32 {
+		self.collection_id
+	}
+
+	fn asset_type(&self) -> u32 {
+		self.asset_type
+	}
+
+	fn dna(&self) -> [u8; 32] {
+		self.dna
+	}
+
+	fn minted_at(&self) -> u32 {
+		self.minted_at
+	}
+}
+
+#[derive(Copy, Clone, Debug, PartialEq, Eq, Encode, Decode, MaxEncodedLen, TypeInfo)]
+pub enum ExampleTransitionId {
+	UpgradeAsset,
+	ConsumeAsset,
+}
+
+/// One specific transition that a game wants to execute.
+pub fn consume_asset(asset: &mut Asset) -> Result<(), sage_api::Error> {
+	if asset.consumed {
+		Err(sage_api::Error::Transition { error: ASSET_ALREADY_CONSUMED })
+	} else {
+		asset.consumed = true;
+		Ok(())
+	}
+}
